Wrapper module to include Cox's package
Cox's package relies on relative paths to run.
It assumes the main Python process is running in its root dir.
This wrapper runs the necessary methods within the root of the submodule
and then moves the working dir back to the original one.
It returns a configured instance of the Migdal class from the submodule
for which the probabilities and total probabilities have already been
computed.